+++ packages/types/src/workspace-config.ts
@@ -126,6 +126,8 @@ export interface ExperimentsConfig {
 	taskOutputBoundaries?: boolean;
 }
 
+export type ExtendsFrom = string | string[];
+
 /** Configures an individual extension. */
 export interface ExtensionConfig {
 	/** Arbitrary configuration that'll be passed to the WASM plugin. */
@@ -371,10 +373,10 @@ export interface WorkspaceConfig {
 	/** Configures experiments across the entire moon workspace. */
 	experiments: ExperimentsConfig;
 	/**
-	 * Extends another workspace configuration file. Supports a relative
+	 * Extends one or many workspace configuration file. Supports a relative
 	 * file path or a secure URL.
 	 */
-	extends: string | null;
+	extends: ExtendsFrom | null;
 	/** Configures extensions that can be executed with `moon ext`. */
 	extensions: Record<string, ExtensionConfig>;
 	/** Configures the generator for scaffolding from templates. */
@@ -748,10 +750,10 @@ export interface PartialWorkspaceConfig {
 	/** Configures experiments across the entire moon workspace. */
 	experiments?: PartialExperimentsConfig | null;
 	/**
-	 * Extends another workspace configuration file. Supports a relative
+	 * Extends one or many workspace configuration file. Supports a relative
 	 * file path or a secure URL.
 	 */
-	extends?: string | null;
+	extends?: ExtendsFrom | null;
 	/** Configures extensions that can be executed with `moon ext`. */
 	extensions?: Record<string, PartialExtensionConfig> | null;
 	/** Configures the generator for scaffolding from templates. */